refactor: 설계 구조에 따라 파일 분리
