+++ src/js/lib/utils/dom.js
@@ -1,2 +1,3 @@
-export const $ = selector => document.querySelector(selector);
-export const $$ = selector => document.querySelectorAll(selector);
+const $ = selector => document.querySelector(selector);
+
+export default $;
