--- src/js/model.js
@@ -1,1 +1,0 @@
-export const lottoData = {};
