+++ src/js/model/lottoData.js
@@ -1,0 +1,4 @@
+export const lottoData = {
+  tickets: [],
+  ticketAmount: 0,
+};
