--- src/js/utils/dom.js
@@ -1,3 +1,0 @@
-const $ = selector => document.querySelector(selector);
-
-export default $;
